+++ turbo_firestore_api/example/lib/turbo_firestore_api/data/dtos/example_dto.dart
@@ -2,7 +2,7 @@ import 'package:turbo_firestore_api/abstracts/turbo_writeable.dart';
 import 'package:turbo_response/turbo_response.dart';
 import 'package:turbo_serializable/models/turbo_serializable_config.dart';
 
-class ExampleDTO extends TurboWriteable {
+class ExampleDTO extends TWriteable {
   ExampleDTO({
     required this.thisIsAString,
     required this.thisIsANumber,