+++ turbo_firestore_api/lib/services/after_sync_t_collection_service.dart
@@ -6,9 +6,9 @@ part of 't_collection_service.dart';
 /// the local state has been updated with new data from Firestore.
 ///
 /// Type Parameters:
-/// - [T] - The document type, must extend [TurboWriteableId<String, void>]
+/// - [T] - The document type, must extend [TWriteableId]
 /// - [API] - The Firestore API type, must extend [TurboFirestoreApi<T>]
-abstract class AfterSyncTCollectionService<T extends TSerializableId<String, void>,
+abstract class AfterSyncTCollectionService<T extends TWriteableId,
     API extends TFirestoreApi<T>> extends TCollectionService<T, API> {
   /// Creates a new [AfterSyncTCollectionService] instance.
   AfterSyncTCollectionService({required super.api});