+++ turbo_firestore_api/lib/services/before_sync_t_document_service.dart
@@ -6,9 +6,9 @@ part of 't_document_service.dart';
 /// the local state is updated with new data from Firestore.
 ///
 /// Type Parameters:
-/// - [T] - The document type, must extend [TurboWriteableId<String, void>]
+/// - [T] - The document type, must extend [TWriteableId]
 /// - [API] - The Firestore API type, must extend [TurboFirestoreApi<T>]
-abstract class BeforeSyncTDocumentService<T extends TSerializableId<String, void>,
+abstract class BeforeSyncTDocumentService<T extends TWriteableId,
     API extends TFirestoreApi<T>> extends TDocumentService<T, API> {
   /// Creates a new [BeforeSyncTDocumentService] instance.
   BeforeSyncTDocumentService({required super.api});